requirements.txt
Flask==0.11.1
gevent==1.1.1
python-memcached==1.58
requests==2.10.0
